Fix compilation error on macOS
Users/runner/work/binlog/binlog/include/binlog/ToStringVisitor.cpp:141:30: error: use of overloaded operator '<<' is ambiguous (with operand types 'binlog::detail::OstreamBuffer' and 'std::size_t' (aka 'unsigned long'))

+++ include/mserialize/Visitor.hpp
@@ -88,14 +88,14 @@ struct Visitor
 
   struct RepeatBegin
   {
-    std::size_t size{}; /**< Number of time the next visted value repeats */
-    string_view tag;    /**< Type tag of the repeating value */
+    std::uint32_t size{}; /**< Number of time the next visted value repeats */
+    string_view tag;      /**< Type tag of the repeating value */
   };
 
   struct RepeatEnd
   {
-    std::size_t size{}; /**< Number of time the previous visted value repeats */
-    string_view tag;    /**< Type tag of the repeating value */
+    std::uint32_t size{}; /**< Number of time the previous visted value repeats */
+    string_view tag;      /**< Type tag of the repeating value */
   };
 
   // visitor methods - to be implemented by derived type
